+++ insights/parsers/x86_debug.py
@@ -5,7 +5,7 @@ Parsers for file ``/sys/kernel/debug/x86/*_enabled`` outputs
 This module provides the following parsers:
 
 X86PTIEnabled - file ``/sys/kernel/debug/x86/pti_enabled``
------------------------------------------------------------
+----------------------------------------------------------
 
 X86IBPBEnabled - file ``/sys/kernel/debug/x86/ibpb_enabled``
 ------------------------------------------------------------
@@ -19,8 +19,26 @@ X86RETPEnabled - file ``/sys/kernel/debug/x86/retp_enabled``
 """
 
 from insights import parser
+from insights import Parser
 from insights.specs import Specs
-from insights.parsers.x86_debug_enabled import X86DebugEnabled
+from insights.parsers import SkipException
+
+
+class X86DebugEnabled(Parser):
+    """
+    Class for parsing file ``/sys/kernel/debug/x86/*_enabled``
+
+    Attributes:
+        value (int): the result parsed of '/sys/kernel/debug/x86/*_enabled'
+
+    Raises:
+        SkipException: When input content is empty
+    """
+    def parse_content(self, content):
+        if not content:
+            raise SkipException("Input content is empty")
+        # it is a digit
+        self.value = int(content[0])
 
 
 @parser(Specs.x86_ibpb_enabled)
@@ -28,7 +46,7 @@ class X86IBPBEnabled(X86DebugEnabled):
     """
     Class for parsing file ``/sys/kernel/debug/x86/ibpb_enabled``
     Typical output of file ``/sys/kernel/debug/x86/retp_enabled`` looks like::
-    1
+        1
 
     Examples:
         >>> type(dva)
@@ -50,7 +68,7 @@ class X86IBRSEnabled(X86DebugEnabled):
     """
     Class for parsing file ``/sys/kernel/debug/x86/ibrs_enabled``
     Typical output of file ``/sys/kernel/debug/x86/ibrs_enabled`` looks like::
-    0
+        0
 
     Examples:
         >>> type(dl)
@@ -72,7 +90,7 @@ class X86PTIEnabled(X86DebugEnabled):
     """
     Class for parsing file ``/sys/kernel/debug/x86/pti_enabled``
     Typical output of file ``/sys/kernel/debug/x86/pti_enabled`` looks like::
-    0
+        0
 
     Examples:
         >>> type(dv)
@@ -94,7 +112,7 @@ class X86RETPEnabled(X86DebugEnabled):
     """
     Class for parsing file ``/sys/kernel/debug/x86/retp_enabled``
     Typical output of file ``/sys/kernel/debug/x86/retp_enabled`` looks like::
-    1
+        1
 
     Examples:
         >>> type(dval)
